# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KKamera API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the signed-in user's password
      description: >-
        Verifies the current password (and a TOTP or backup code when 2FA is
        enabled), sets the new password, and signs out every other session.
        Returns a fresh token so the calling device stays signed in. Errors are
        400 (never 401) so a mistyped password doesn't end the session.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed; new session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input, wrong current password, or invalid 2FA code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/2fa/setup:
    post:
      operationId: setup2FA
      tags: [auth]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 2FA setup info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFASetup"
  /auth/2fa/verify:
    post:
      operationId: verify2FA
      tags: [auth]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFAVerifyInput"
      responses:
        "200":
          description: 2FA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/2fa/disable:
    post:
      operationId: disable2FA
      tags: [auth]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFAVerifyInput"
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        totpCode:
          type: ["string", "null"]
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          description: At least 8 characters and at most 72 bytes (UTF-8).
        totpCode:
          type: ["string", "null"]
          description: Required when 2FA is enabled — a 6-digit TOTP or a backup code.
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, email, name, referralCode, twoFAEnabled, onboardingCompleted, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        referralCode:
          type: string
        twoFAEnabled:
          type: boolean
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
...
    TwoFASetup:
      type: object
      required: [secret, qrCodeUrl, backupCodes]
      properties:
        secret:
          type: string
        qrCodeUrl:
          type: string
        backupCodes:
          type: array
          items:
            type: string
    TwoFAVerifyInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
...